// resources/js/types/AreaFormValues.ts
export type AreaFormValues = {
    id_area: string;
    name: string;
    nomenclatura: string;
    id_entidad_federativa: string;
    municipio_id: string;
    codigo_postal: string;
    colonia: string;
    calle: string;
    numero_exterior: string;
    numero_interior: string;
    telefono: string;
};